Add files via upload
The databases have been taken directly from use as a copy added here, so they already contain the changes I mentioned that are supported by the source code of the website. but contains the server and some ready-made values, e.g. account and login account and a pre-created character and clan. take these things into account if you decide to use the GunzDB database, while the Gunzweb database is a mandatory add-on for the site's functionality.